Split up test/all.js into separate files
This allows jest to split up test runs.
We can now have one test file for each source file.

(cherry picked from commit 92b3084e1800896634c32c16fd647a74c6f7f840)
